backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
redis==5.0.1
firebase-admin==6.4.0
email-validator==2.1.0
websockets==12.0
